rename dir, rationalise imports __init__
